backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.3.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
scikit-image>=0.23.0
ultralytics>=8.2.0
groq>=0.9.0
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
